feat: implement advanced analysis worker
Make display a label-only
